app updated

+++ fizzbuzz/src/main/java/org/dex/fizzbuzz/entity/FizzBuzValue.java
@@ -1,12 +1,18 @@
 package org.dex.fizzbuzz.entity;
 
+import javax.validation.constraints.Max;
+import javax.validation.constraints.Min;
+
+import org.hibernate.validator.constraints.Range;
+
 import lombok.Data;
 
 @Data
 public class FizzBuzValue {
 	
+   
 	
-	public long in;
+	public int in;
 	
 	public String result;
